refactor: CarName 원시값 포장 및 생성자 오버로딩, 자동차 이름 입력 검증 책임 RacingManager에게 옮김
