+++ src/main/java/racingcar/domain/Cars.java
@@ -6,16 +6,16 @@ import java.util.stream.Stream;
 import racingcar.domain.numbergenerator.NumberGenerator;
 
 public class Cars {
-    private static final String INPUT_NAMES_REGEX = "^[A-Za-z0-9,]*[A-Za-z0-9]$";
-    private static final String NAME_DELIMITER = ",";
     private final List<Car> cars;
 
-    public Cars(String inputNames, NumberGenerator numberGenerator) {
-        validateInputNames(inputNames);
-        String[] names = inputNames.split(NAME_DELIMITER);
-        cars = Arrays.stream(names)
+    public Cars(List<CarName> carNames, NumberGenerator numberGenerator) {
+        cars = carNames.stream()
                 .map(name -> new Car(name, numberGenerator))
                 .collect(Collectors.toList());
+    }
+
+    public Cars(String inputNames, NumberGenerator numberGenerator) {
+        this(CarName.parseFrom(inputNames), numberGenerator);
     }
 
     public void allTryMove() {
@@ -24,11 +24,5 @@ public class Cars {
 
     public List<Car> getCars() {
         return Collections.unmodifiableList(cars);
-    }
-
-    private void validateInputNames(String inputNames) {
-        if (!inputNames.matches(INPUT_NAMES_REGEX)) {
-            throw new IllegalArgumentException();
-        }
     }
 }
